Fix MedicineAvailabilityResponse quantity type to float
Database has decimal quantities (e.g. 8.767), int type caused
pydantic validation error.

+++ backend/app/api/staff.py
@@ -118,7 +118,7 @@ class MedicineAvailabilityResponse(BaseModel):
     pharmacy_name: str
     is_available: bool
     price: float | None = None
-    quantity: int | None = None
+    quantity: float | None = None
 
 
 class MedicineResponse(BaseModel):
